+++ NEWS.md
@@ -25,6 +25,13 @@
 -   Updated sparse matrix conversion routines for compatibility with
     scipy 1.11.0.
 
+-   New function `virtualenv_starter()`, which can be used to find a suitable
+    python binary for creating a virtual environmnent. This is now the default
+    method for finding the python binary when calling `virtualenv_create(version = <version>)`.
+
+
+-   `install_python()` now gives a better error message if git is not installed.
+
 -   Fixed an issue where a py capsule finalizer could access the R API from
     a background thread. (#1406)